feat(bootstrap): allow local archive reuse from bootstrap

+++ bootstrap.ps1
@@ -1,5 +1,5 @@
 param(
-    [Parameter(Mandatory=$false,HelpMessage="Try to find a local tollset.zip to install from")][string]$local=$true
+    [Parameter(Mandatory=$false,HelpMessage="Try to find a local toolset.zip to install from")][string]$local=$true
 )
 try {
     Set-StrictMode -Version Latest
@@ -12,17 +12,32 @@ try {
     Write-Host "+--------------------------------+" -ForegroundColor Cyan
 
     # Download archive
-
-    Write-Output "About to download toolset..."
-    $url="https://github.com/ETML-INF/standard-toolset/releases/latest/download/toolset.zip"
+    if($local)
+    {
+	$localarchivepath = (Get-ChildItem -Path ".\toolset*.zip" | Select-Object -First 1).FullName
+	if(![string]::IsNullOrEmpty($localarchivepath))
+	{
+	    $archivepath = $localarchivepath
+	    Write-Host "Found local $archivepath"
+	}
+	else{
+	    Write-Error "No local archive found, aborting local install"
+	    Exit 2
+	}
+    }
+    else{
+	Write-Output "About to download toolset..."
+	$url="https://github.com/ETML-INF/standard-toolset/releases/latest/download/toolset.zip"
+	$timestamp = Get-Date -format yyyy_MM_dd_H_mm_ss
+	$archivename = "toolset-$timestamp"
+	$archivepath = "$env:TEMP\$archivename.zip"
+	Invoke-WebRequest -Uri "$url" -OutFile "$archivepath"
+	
+    }
     
-    $timestamp = Get-Date -format yyyy_MM_dd_H_mm_ss
-    $archivename = "toolset-$timestamp"
-    $archivepath = "$env:TEMP\$archivename.zip"
-    Invoke-WebRequest -Uri "$url" -OutFile "$archivepath"
-
+    
     # Extract
-    $archivedirectory = "$env:TEMP\$archivename"
+    $archivedirectory = "$env:TEMP\toolset-$(New-Guid)"
     Write-Output "About to extract $archivepath to $archivedirectory"
     Expand-Archive $archivepath $archivedirectory
 
@@ -33,7 +48,11 @@ try {
 
     # Cleaning up
     Remove-Item $archivedirectory
-    Remove-Item $archivepath
+    if(!$local)
+    {
+	Remove-Item $archivepath	
+    }
+    
 }
 catch {
     Write-Error "Something went wrong: $_. Please contact the maintainer for more info..."
